backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
alembic==1.13.3
psycopg2-binary==2.9.10
supabase==2.10.0
pydantic==2.9.2
python-dotenv==1.0.1
python-multipart==0.0.12
anthropic>=0.40.0
pdfminer.six==20250506
python-docx==1.1.2
google-api-python-client==2.149.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
